Added NotificationEvents directly to MonitoredStocks Made app functionality entirely account based

+++ src/test/java/com/jruchel/stockmonitor/services/stocks/MonitoredStockServiceTest.java
@@ -19,9 +19,11 @@ import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -58,12 +60,13 @@ class MonitoredStockServiceTest {
 
     @BeforeEach
     void beforeEach() {
-        userRepository.deleteAll();
-        roleRepository.deleteAll();
+        if (userRepository.count() > 0) userRepository.deleteAll();
+        if (roleRepository.count() > 0) roleRepository.deleteAll();
         Role userRole = roleService.createUserRole();
         user = userService.save(getTestUser(userRole));
         Mockito.when(securityService.getCurrentUser()).thenReturn(user);
     }
+
 
     private User getTestUser(Role role) {
         return User.builder().monitoredStocks(new ArrayList<>()).username("test").password("test").email("test@test.com").roles(Set.of(role)).build();
